refactor(StringCalculator): 식 변환기 이름 변경
