+++ src/main/kotlin/stringcalculator/NumberTokenizer.kt
@@ -1,6 +1,6 @@
 package stringcalculator
 
-import stringcalculator.ExpressionAnalyzer.Companion.Expression
+import stringcalculator.ExpressionMatcher.Companion.Expression
 
 class NumberTokenizer {
