+++ src/test/kotlin/stringcalculator/ExpressionMatcherTest.kt
@@ -3,11 +3,11 @@ package stringcalculator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import stringcalculator.ExpressionAnalyzer.Companion.Expression
+import stringcalculator.ExpressionMatcher.Companion.Expression
 
-internal class ExpressionAnalyzerTest {
+internal class ExpressionMatcherTest {
 
-    private val sut = ExpressionAnalyzer()
+    private val sut = ExpressionMatcher()
 
     @DisplayName("커스텀 구분자가 없을경우 기본 구분자로 쉼표(,)와 콜론(:)을 포함시킨다")
     @Test
@@ -16,7 +16,7 @@ internal class ExpressionAnalyzerTest {
         val expressionInput = "1:2,3"
 
         // when
-        val actual = sut.analyze(expressionInput)
+        val actual = sut.transform(expressionInput)
 
         // then
         assertThat(actual).isEqualTo(Expression(DEFAULT_DELIMITERS, expressionInput))
@@ -29,7 +29,7 @@ internal class ExpressionAnalyzerTest {
         val expressionInput = "//;\n1:2,3;4"
 
         // when
-        val actual = sut.analyze(expressionInput)
+        val actual = sut.transform(expressionInput)
 
         // then
         assertThat(actual).isEqualTo(Expression(listOf(*DEFAULT_DELIMITERS.toTypedArray(), ";"), "1:2,3;4"))
@@ -42,7 +42,7 @@ internal class ExpressionAnalyzerTest {
         val expressionInput = ",a;2:3,12"
 
         // when
-        val actual = sut.analyze(expressionInput)
+        val actual = sut.transform(expressionInput)
 
         // then
         assertThat(actual).isEqualTo(Expression(DEFAULT_DELIMITERS, ",a;2:3,12"))
