+++ src/test/kotlin/stringcalculator/NumberTokenizerTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import stringcalculator.ExpressionAnalyzer.Companion.Expression
+import stringcalculator.ExpressionMatcher.Companion.Expression
 
 internal class NumberTokenizerTest {
